# storage/api-docs/api-docs.yaml
openapi: 3.0.0
info:
  title: 'QMS API Documentation'
  description: 'Secure messenger API with end-to-end encryption'
  contact:
    name: Heruvim6w
    email: heruvim.6w@gmail.com
  license:
    name: MIT
    url: 'https://opensource.org/licenses/MIT'
  version: 1.0.0
servers:
  -
    url: 'http://localhost:8000'
    description: 'QMS API Server'
  -
    url: 'http://localhost:8000'
    description: 'Development Server'
    variables:
      protocol:
        enum:
          - http
          - https
        default: http
paths:
...
  /api/v1/messages:
    get:
      tags:
        - Messages
      summary: 'Get all messages in a chat'
      description: 'Retrieve all messages from a specific chat with pagination. Messages are automatically decrypted. Only accessible to chat members. Messages are sorted by creation date (oldest first).'
      operationId: getMessages
      parameters:
        -
          name: chat_id
          in: query
          description: 'ID of the chat'
          required: true
          schema:
            type: integer
            example: 1
        -
          name: limit
          in: query
          description: 'Maximum number of messages to return (default: 50)'
          schema:
            type: integer
            example: 50
        -
          name: offset
          in: query
          description: 'Number of messages to skip (for pagination)'
          schema:
            type: integer
            example: 0
      responses:
        200:
          description: 'List of messages in chat (decrypted)'
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/MessageResponse'
        400:
          description: 'Missing chat_id parameter'
        403:
          description: 'Access denied - user not a member of this chat'
      security:
        -
          bearerAuth: []
    post:
      tags:
        - Messages
      summary: 'Send an encrypted message'
      description: 'Send a message to a chat. Message is end-to-end encrypted. Either existing chat_id or new receiver_id should be provided. If receiver_id is provided, private chat is created automatically.'
      operationId: sendMessage
      requestBody:
        description: 'Message data'
        required: true
        content:
          application/json:
            schema:
              required:
                - content
              properties:
                chat_id:
                  description: 'ID of existing chat (if empty, receiver_id must be provided)'
                  type: integer
                  example: 1
                  nullable: true
                receiver_id:
                  description: 'UUID of receiver (for new private chats, if empty chat_id must be provided)'
                  type: string
                  format: uuid
                  example: 550e8400-e29b-41d4-a716-446655440001
                  nullable: true
                content:
                  description: 'Message text (encrypted before sending)'
                  type: string
                  maxLength: 65535
                  example: 'Hello, how are you?'
                type:
                  description: 'Message type (default: text)'
                  type: string
                  enum:
                    - text
                    - image
                    - voice
                    - video
                    - file
                  example: text
              type: object
      responses:
        201:
          description: 'Message sent successfully'
          content:
            application/json:
              schema:
                properties:
                  id:
                    type: integer
                    example: 1
                  chat_id:
                    type: integer
                    example: 1
                  status:
                    type: string
                    example: sent
                  created_at:
                    type: string
                    format: date-time
                type: object
        403:
          description: 'Access denied - not a member of this chat'
        404:
          description: 'Chat or receiver not found'
        422:
          description: 'Validation error'
      security:
        -
          bearerAuth: []
  '/api/v1/messages/{id}':
    get:
      tags:
        - Messages
      summary: 'Get a specific message'
      description: 'Retrieve and decrypt a specific message by ID. Message is automatically marked as read. Only accessible to chat members.'
      operationId: getMessage
      parameters:
        -
          name: id
          in: path
          description: 'Message ID'
          required: true
          schema:
            type: integer
            example: 1
      responses:
        200:
          description: 'Message details (decrypted)'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse_v2'
        403:
          description: 'Access denied - user not a member of this chat'
        404:
          description: 'Message not found'
        500:
          description: 'Decryption error'
      security:
        -
          bearerAuth: []
  '/api/v1/messages/{id}/upload':
    post:
      tags:
        - Messages
      summary: 'Upload a file attachment to a message'
      description: 'Attach a file to a message. File is stored encrypted on server. Only message author can add attachments. Supports any file type within size limits.'
      operationId: uploadFileToMessage
      parameters:
        -
          name: id
          in: path
          description: 'Message ID'
          required: true
          schema:
            type: integer
            example: 1
      requestBody:
        description: 'File to upload'
        required: true
        content:
          multipart/form-data:
            schema:
              required:
                - file
              properties:
                file:
                  description: 'Binary file content (max 100MB recommended)'
                  type: string
                  format: binary
              type: object
      responses:
        200:
          description: 'File uploaded and attached successfully'
          content:
            application/json:
              schema:
                properties:
                  status:
                    type: string
                    example: file_uploaded
                  attachment_id:
                    type: integer
                    example: 1
                  file_path:
                    type: string
                    example: uploads/messages/file.jpg
                  file_size:
                    type: integer
                    example: 102400
                  mime_type:
                    type: string
                    example: image/jpeg
                type: object
        403:
          description: 'Access denied - not the message author'
        404:
          description: 'Message not found'
        413:
          description: 'File too large'
        422:
          description: 'Validation error (invalid file type, etc)'
      security:
        -
          bearerAuth: []
...
components:
  schemas:
...
    UserProfileResponse:
      title: 'User Profile'
      description: 'Complete user profile information'
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        email:
          type: string
          format: email
        uin:
          description: '8-digit unique identifier'
          type: string
        username:
          description: 'Custom username (3-20 chars, latin only)'
          type: string
          nullable: true
        status:
          description: 'Current online status'
          type: string
          enum:
            - online
            - offline
        online_status:
          description: 'Selected status mood'
          type: string
          enum:
            - online
            - chatty
            - angry
            - depressed
            - home
            - work
            - eating
            - away
            - unavailable
            - busy
            - do_not_disturb
        custom_status:
          description: 'User custom status text with emoji support'
          type: string
          maxLength: 50
          nullable: true
        last_seen_at:
          description: 'Last online timestamp'
          type: string
          format: date-time
          nullable: true
        locale:
          description: 'User interface language'
          type: string
          example: ru
        created_at:
          type: string
          format: date-time
      type: object
...
    MessageResponse:
      title: Message
      description: 'Decrypted message object'
      properties:
        id:
          type: integer
        chat_id:
          type: integer
        sender:
          properties:
            id:
              type: string
              format: uuid
            name:
              type: string
            uin:
              type: string
            username:
              type: string
              nullable: true
          type: object
        content:
          description: 'Decrypted message text'
          type: string
        type:
          type: string
          enum:
            - text
            - image
            - voice
            - video
            - file
        attachments:
          type: array
          items:
            $ref: '#/components/schemas/AttachmentResponse'
        is_read:
          type: boolean
        created_at:
          type: string
          format: date-time
      type: object
    AttachmentResponse:
      title: Attachment
      description: 'File attachment'
      properties:
        id:
          type: integer
        message_id:
          type: integer
        file_path:
          type: string
        mime_type:
          type: string
          example: image/jpeg
        size:
          description: 'File size in bytes'
          type: integer
        name:
          description: 'Original filename'
          type: string
        created_at:
          type: string
          format: date-time
      type: object
...
    MessageResponse_v2:
      title: 'Message Response v2'
      description: 'Complete message response with metadata'
      properties:
        id:
          type: integer
        chat_id:
          type: integer
        sender:
          $ref: '#/components/schemas/UserProfileResponse'
        content:
          type: string
        type:
          type: string
          enum:
            - text
            - image
            - voice
            - video
            - file
        attachments:
          type: array
          items:
            $ref: '#/components/schemas/AttachmentResponse'
        read_by:
          description: 'List of users who read this message'
          type: array
          items:
            properties:
              id:
                type: string
                format: uuid
              name:
                type: string
              read_at:
                type: string
                format: date-time
            type: object
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
      type: object
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
security:
  -
    bearerAuth: []